New Foreign Function Interface (FFI)
This interface makes the types of imported/exported
functions independent of the particular representation
choices made by the runtime system.

Instead, values of arguments and results are converted
between the type used for ASL code and a standard
representation used for C code.

The representation of types is as follows

- if the ASL type is boolean then "bool" (from stdbool.h)
- if the ASL type is string then "const char*"
- if the ASL type is integer then "int"
- if the ASL type is bits(N) then either
  - uint8_t, uint16_t, uint32_t, uint64_t if N IN {8, 16, 32, 64}
  - array uint64_t _[N'] where N' is ceiling(N / 64)
- if the ASL type is an enumeration, then that enumeration
  provided that the enumeration type has been explicitly
  exported

Other types cannot be used in imported/exported functions.

In addition, exported functions must not raise exceptions
and, for extra safety, we insert a runtime check that
they do not raise an exception

In addition, for each configuration variable 'v : ty', we generate
a pair of getter/setter functions

    ASL_set_config_{v}({cty} value);
    {cty} ASL_get_config_{v}();

Note that (based on the ASL semantics of configuration variables)
all changes to the configuration variable should be made
before executing any ASL code.

Signed-off-by: Alastair Reid <alastair.reid@intel.com>