Tests: stop using %decolor
We used to have a problem in tests that checked error messages
because error messages were displayed using color and
the colorization inserted terminal escape characters into the output.

To handle this, we filtered error message output through a sed script.
But the sed script was not entirely portable.
And we had to use %decolor on lots of tests.

Now that we disable color when in batchmode, this whole mess is
not necessary.