Runtime: print error messages to stdout
Printing error messages to stderr makes testing more complex
because it is hard to portably redirect stderr and stdout
is not synchronized with stderr.
So it is much easier just to use stdout.

Note that we expect most users to override the definitions
of ASL_{runtime_error,assert,error} and they may use
stderr in their own versions if they choose.