Github workflow: use older ubuntu
With ubuntu-latest, we can no longer install clang-16 using the llvm.sh script.
It seems that the llvm.sh script broke when 22.04 was first released so maybe
this is a temporary blip because ubuntu-latest just changed to 22.04.1.

This crude workaround just uses an older version of Ubuntu - hopefully fixing
the problem until we are ready to change to using clang-17 or later.

Signed-off-by: Alastair Reid <alastair.reid@intel.com>