Cleanup: generate prototypes
In the past, we treated the Prelude differently from user-written
code by not generating function prototypes for functions defined in
the Prelude.

This restriction is no longer needed and it was causing problems in
the backend where we could not rely on the existence of a function
prototype for every function.

This change ensures that we generate function prototypes for all functions
that have a FunDefn.

Signed-off-by: Alastair Reid <alastair.reid@intel.com>