Exception checks: avoid false positives
The original exception check was a bit approximate.
It would report that a function could throw an exception even if
the function caught all exceptions safely.

The newer exception check that is based on exception markers
is much more accurate and avoids those false positives.

This change disables the original check when we are using
the new exception check to avoid those false positives.

Signed-off-by: Alastair Reid <alastair.reid@intel.com>