Remove comment tracking
Some time ago, we added the option of tracking comments so
that the pretty-printer could insert comments when displaying source
code.

This has not been used for some time and was always going to be
inferior to extracting code from the original input file.
Removing all this code to reduce clutter.