Language extension: Print
This adds a polymorphic print function inspired by Python f-strings.

e.g., 'Print(x={x}n);' prints 'x=3\n' (if x = 3).

At the moment, formatting is limited to just printing the value with no
format control and the types of variables printed is limited to
bitvector, integer, string or boolean.

The plan is to support user-defined types and control over decimal/hex,
etc. based on a subset of Python f-strings.

Signed-off-by: Alastair Reid <alastair.reid@intel.com>