Xform_bitslices
Fixes a bug that occurs if you have code like this

    let wide : bits(1024) = ...;
    let r = [wide[0 +: 64], Zeros(64)];

The bug was causing us to generate a zero_extend from 1024 bits "up" to
64 bits.

This patch does not fully fix the bug but it prevents it from happening
when the width of the bitslice is a constant - which fixes the problem
case.

The most general fix would probably be to find the maximum width that
appears anywhere in the transformed expression, widen everything to that
maximum width and then use and/or/mask/etc. in the transformation.
Leaving that to another day.